[fix] 피드백 반영

+++ src/main/java/step2/domain/lottoResult/LottoResult.java
@@ -21,7 +21,7 @@ public class LottoResult {
         return winCount.moneyCount(winNumber.getMoney());
     }
 
-    public void win(long hitNumber) {
+    public void win(int hitNumber) {
         if (winNumber.getHit() != hitNumber) {
             return;
         }
